TST: Split test_excel.py into sub test files #24472
